+++ packages/keychain/CHANGELOG.md
@@ -47,6 +47,41 @@
   * devDependencies
     * @libp2p/peer-id-factory bumped from ^4.0.3 to ^4.0.4
 
+## [5.2.8](https://github.com/libp2p/js-libp2p/compare/keychain-v5.2.7...keychain-v5.2.8) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/utils bumped from ^6.7.0 to ^6.7.1
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+
+## [5.2.7](https://github.com/libp2p/js-libp2p/compare/keychain-v5.2.6...keychain-v5.2.7) (2025-06-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/utils bumped from ^6.6.7 to ^6.7.0
+
+## [5.2.6](https://github.com/libp2p/js-libp2p/compare/keychain-v5.2.5...keychain-v5.2.6) (2025-06-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.5 to ^5.1.6
+    * @libp2p/interface bumped from ^2.10.3 to ^2.10.4
+    * @libp2p/utils bumped from ^6.6.6 to ^6.6.7
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.19 to ^5.1.20
+
 ## [5.2.5](https://github.com/libp2p/js-libp2p/compare/keychain-v5.2.4...keychain-v5.2.5) (2025-06-03)