+++ packages/metrics-opentelemetry/CHANGELOG.md
@@ -1,5 +1,38 @@
 # Changelog
 
+## [1.0.20](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.19...opentelemetry-metrics-v1.0.20) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/utils bumped from ^6.7.0 to ^6.7.1
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+
+## [1.0.19](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.18...opentelemetry-metrics-v1.0.19) (2025-06-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/utils bumped from ^6.6.7 to ^6.7.0
+
+## [1.0.18](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.17...opentelemetry-metrics-v1.0.18) (2025-06-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.3 to ^2.10.4
+    * @libp2p/utils bumped from ^6.6.6 to ^6.6.7
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.19 to ^5.1.20
+
 ## [1.0.17](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.16...opentelemetry-metrics-v1.0.17) (2025-06-03)