+++ packages/protocol-dcutr/CHANGELOG.md
@@ -44,6 +44,43 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [2.0.36](https://github.com/libp2p/js-libp2p/compare/dcutr-v2.0.35...dcutr-v2.0.36) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/interface-internal bumped from ^2.3.17 to ^2.3.18
+    * @libp2p/utils bumped from ^6.7.0 to ^6.7.1
+
+## [2.0.35](https://github.com/libp2p/js-libp2p/compare/dcutr-v2.0.34...dcutr-v2.0.35) (2025-06-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^2.3.16 to ^2.3.17
+    * @libp2p/utils bumped from ^6.6.7 to ^6.7.0
+
+## [2.0.34](https://github.com/libp2p/js-libp2p/compare/dcutr-v2.0.33...dcutr-v2.0.34) (2025-06-16)
+
+
+### Bug Fixes
+
+* update multiaddr ([#3184](https://github.com/libp2p/js-libp2p/issues/3184)) ([6c42ea6](https://github.com/libp2p/js-libp2p/commit/6c42ea64a6e22028a87ecb3422e418e99ff09279))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.3 to ^2.10.4
+    * @libp2p/interface-internal bumped from ^2.3.15 to ^2.3.16
+    * @libp2p/utils bumped from ^6.6.6 to ^6.6.7
+
 ## [2.0.33](https://github.com/libp2p/js-libp2p/compare/dcutr-v2.0.32...dcutr-v2.0.33) (2025-06-03)