+++ packages/pubsub-floodsub/CHANGELOG.md
@@ -98,6 +98,64 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [10.1.42](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.41...floodsub-v10.1.42) (2025-06-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.4 to ^2.10.5
+    * @libp2p/pubsub bumped from ^10.1.16 to ^10.1.17
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/interface-compliance-tests bumped from ^6.4.12 to ^6.4.13
+    * @libp2p/logger bumped from ^5.1.20 to ^5.1.21
+    * @libp2p/peer-collections bumped from ^6.0.33 to ^6.0.34
+    * @libp2p/peer-id bumped from ^5.1.7 to ^5.1.8
+
+## [10.1.41](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.40...floodsub-v10.1.41) (2025-06-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.11 to ^6.4.12
+
+## [10.1.40](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.39...floodsub-v10.1.40) (2025-06-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/pubsub bumped from ^10.1.15 to ^10.1.16
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.10 to ^6.4.11
+    * @libp2p/peer-collections bumped from ^6.0.32 to ^6.0.33
+
+## [10.1.39](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.38...floodsub-v10.1.39) (2025-06-16)
+
+
+### Bug Fixes
+
+* update multiaddr ([#3184](https://github.com/libp2p/js-libp2p/issues/3184)) ([6c42ea6](https://github.com/libp2p/js-libp2p/commit/6c42ea64a6e22028a87ecb3422e418e99ff09279))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.10.3 to ^2.10.4
+    * @libp2p/pubsub bumped from ^10.1.14 to ^10.1.15
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.5 to ^5.1.6
+    * @libp2p/interface-compliance-tests bumped from ^6.4.9 to ^6.4.10
+    * @libp2p/logger bumped from ^5.1.19 to ^5.1.20
+    * @libp2p/peer-collections bumped from ^6.0.31 to ^6.0.32
+    * @libp2p/peer-id bumped from ^5.1.6 to ^5.1.7
+
 ## [10.1.38](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.37...floodsub-v10.1.38) (2025-06-03)